feat(aigc): 完善AI生成内容管理和文章发布功能
- 将AI生成内容的isUsed字段从Integer改为Long类型
- 添加AI生成内容的isUsed字段验证和导出功能
- 更新AI生成内容查询逻辑以支持isUsed条件查询
- 添加文章任务明细的回调接口和根据主任务查询详情功能
- 扩展文章任务明细实体增加发布截图、控评、比特相关字段
- 实现文章详情页关联RPA账号名称显示功能
- 添加终端类型字段到文章任务相关实体
- 在系统配置中添加文章任务和回调接口路径白名单

+++ ruoyi-modules/aigc/src/main/java/org/dromara/rp/domain/bo/RpAccountGroupBo.java
@@ -46,7 +46,7 @@ public class RpAccountGroupBo extends BaseDo {
         /**
          * 关联rpa账号
          */
-        private Integer rpaNo;
+        private Long rpaNo;
 
 
 }
